move queries inside fetch hook
